[Fatih][Kaplan]
I have created a class named GPSTracker which collects the coordinates of current location and also updates it in a given interval of time and position change.I have set the latitude and longitude in Helper class.Just for the test I have done a double to float operation but not in a great coding style.By the way, our Position class requires lat and lon values in Float, but location manager returns variables in Double.Will we continue this way or change the variable types?

I have also created a GPSActivity class in case I might wanna do some things in it.